+++ android/src/main/kotlin/com/deepanshuchaudhary/pdf_bitmaps/PdfBitmap.kt
@@ -6,7 +6,6 @@ import android.graphics.Color
 import android.graphics.Matrix
 import android.graphics.Rect
 import android.graphics.pdf.PdfRenderer
-import android.net.Uri
 import android.os.Build
 import android.os.ParcelFileDescriptor
 import android.util.Log
@@ -14,7 +13,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlinx.coroutines.yield
 import java.io.ByteArrayOutputStream
-import java.io.File
 import java.io.FileNotFoundException
 import kotlin.math.floor
 
@@ -22,13 +20,18 @@ import kotlin.math.floor
 // https://developer.android.com/reference/android/graphics/pdf/PdfRenderer
 // https://stackoverflow.com/questions/2883355/how-to-render-pdf-in-android
 
+data class PageInfo(
+    var pageNumber: Int,
+    val rotationAngle: Int,
+    val scale: Double,
+    val backgroundColor: String?
+)
+
 // For getting pdf file page count.
 suspend fun getPdfBitmap(
     pdfPath: String?,
     context: Activity,
-    pagesIndexes: List<Int>,
-    scale: Double,
-    backgroundColor: String?
+    pagesInfo: List<PageInfo>,
 ): List<ByteArray>? {
 
     var byteArrayList: MutableList<ByteArray> = mutableListOf()
@@ -41,7 +44,7 @@ suspend fun getPdfBitmap(
 
         suspend fun renderPage(
             parcelFileDescriptor: ParcelFileDescriptor?,
-            pagesIndexes: List<Int>
+            pagesInfo: List<PageInfo>
         ) {
             try {
                 yield()
@@ -49,27 +52,28 @@ suspend fun getPdfBitmap(
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && parcelFileDescriptor != null) {
                     pdfRenderer = PdfRenderer(parcelFileDescriptor)
 
-                    val pdfBackgroundColor = try {
-                        if (backgroundColor != null)
-                            Color.parseColor(backgroundColor)
-                        else
-                            Color.TRANSPARENT
-                    } catch (e: Exception) {
-                        Log.d(PdfBitmapsPlugin.LOG_TAG, "createPdfBitmaps - IN")
-                        Log.e("Parse", "Error parsing $backgroundColor. $e")
-                        Color.TRANSPARENT
-                    }
-
 //                    var pageCount: Int = pdfRenderer.pageCount
 
-                    pagesIndexes.forEach { pageIndex ->
+                    pagesInfo.forEach { pageInfo ->
                         yield()
-                        val page: PdfRenderer.Page = pdfRenderer.openPage(pageIndex)
+
+                        val pdfBackgroundColor = try {
+                            if (pageInfo.backgroundColor != null)
+                                Color.parseColor(pageInfo.backgroundColor)
+                            else
+                                Color.TRANSPARENT
+                        } catch (e: Exception) {
+                            Log.d(PdfBitmapsPlugin.LOG_TAG, "createPdfBitmaps - IN")
+                            Log.e("Parse", "Error parsing ${pageInfo.backgroundColor}. $e")
+                            Color.TRANSPARENT
+                        }
+
+                        val page: PdfRenderer.Page = pdfRenderer.openPage(pageInfo.pageNumber - 1)
 
 //                      val width: Int = context.resources.displayMetrics.densityDpi / 72 * page.width
 
-                        val width: Int = floor(page.width * scale.toFloat()).toInt()
-                        val height: Int = floor(page.height * scale.toFloat()).toInt()
+                        val width: Int = floor(page.width * pageInfo.scale.toFloat()).toInt()
+                        val height: Int = floor(page.height * pageInfo.scale.toFloat()).toInt()
                         val bitmap: Bitmap = Bitmap.createBitmap(
                             width, height,
                             Bitmap.Config.ARGB_8888
@@ -80,27 +84,31 @@ suspend fun getPdfBitmap(
                         val matrix = Matrix()
                         matrix.postTranslate((-0).toFloat(), (-0).toFloat())
 
-                        if (scale.toFloat() != 1.0f)
-                            matrix.postScale(scale.toFloat(), scale.toFloat())
+                        if (pageInfo.scale.toFloat() != 1.0f)
+                            matrix.postScale(pageInfo.scale.toFloat(), pageInfo.scale.toFloat())
 
                         page.render(
                             bitmap,
                             Rect(
                                 0,
                                 0,
-                                floor(page.width * scale.toFloat()).toInt(),
-                                floor(page.height * scale.toFloat()).toInt()
+                                floor(page.width * pageInfo.scale.toFloat()).toInt(),
+                                floor(page.height * pageInfo.scale.toFloat()).toInt()
                             ),
-                            null,
+                            matrix,
                             PdfRenderer.Page.RENDER_MODE_FOR_DISPLAY
                         )
                         page.close()
 
                         val stream = ByteArrayOutputStream()
-                        bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream)
+
+                        val rotatedBitmap =
+                            Utils().rotateBitmap(bitmap, pageInfo.rotationAngle.toFloat())
+                        bitmap.recycle()
+                        rotatedBitmap.compress(Bitmap.CompressFormat.PNG, 100, stream)
                         var byteArray: ByteArray = stream.toByteArray()
                         byteArrayList.add(byteArray)
-                        bitmap.recycle()
+                        rotatedBitmap.recycle()
                         stream.close()
                     }
 
@@ -116,7 +124,7 @@ suspend fun getPdfBitmap(
         val parcelFileDescriptor: ParcelFileDescriptor? =
             contentResolver.openFileDescriptor(Utils().getURI(pdfPath!!), "r")
 
-        renderPage(parcelFileDescriptor, pagesIndexes)
+        renderPage(parcelFileDescriptor, pagesInfo)
 
         val end = System.nanoTime()
         println("Elapsed time in nanoseconds: ${end - begin}")