+++ android/src/main/kotlin/com/deepanshuchaudhary/pdf_bitmaps/Utils.kt
@@ -1,7 +1,9 @@
 package com.deepanshuchaudhary.pdf_bitmaps
 
+import android.graphics.Bitmap
+import android.graphics.Matrix
 import android.net.Uri
-import java.io.*
+import java.io.File
 
 class Utils {
     fun getURI(uri: String): Uri {
@@ -10,5 +12,11 @@ class Utils {
         return if ((parsedScheme == null) || parsedScheme.isEmpty()) {
             Uri.fromFile(File(uri))
         } else parsed
+    }
+
+    fun rotateBitmap(source: Bitmap, angle: Float): Bitmap {
+        val matrix = Matrix()
+        matrix.postRotate(angle)
+        return Bitmap.createBitmap(source, 0, 0, source.width, source.height, matrix, true)
     }
 }