+++ lib/pdf_bitmaps.dart
@@ -1,4 +1,4 @@
 export 'src/pdf_bitmaps.dart' show PdfBitmaps;
 
 export 'src/pdf_bitmaps_method_channel.dart'
-    show PDFBitmapParams, PDFPageCountParams, PDFBitmapsParams;
+    show PDFBitmapParams, PDFPageCountParams, PDFBitmapsParams, PageInfo;